chore: Refactor chips
BREAKING CHANGE: `Material.Choice.set` was renamed to
`Material.ChipSet.Choice.chipSet`. The same is true for filter and input
chips.

BREAKING CHANGE: Choice chip API changed.

Before:

```elm
import Material.Chip.Choice as ChoiceChip

ChoiceChip.set []
    [ ChoiceChip.chip
          (ChoiceChip.config
              |> ChoiceChip.setSelected (Just True)
              |> ChoiceChip.setOnClick (Clicked "Chip")
          )
          "Chip"
    ]
```

After:

```elm
import Material.Chip.Choice as ChoiceChip
improt Material.ChipSet.Choice as ChoiceChipSet

ChoiceChipSet.chipSet
    (ChoiceChipSet.config { toLabel = identity }
        |> ChoiceChipSet.setSelected (Just "Chip")
        |> ChoiceChipSet.setOnChange Clicked
    )
    [ ChoiceChip.chip ChoiceChip.config "Chip" ]
```

+++ bin/make-patches.sh
@@ -10,10 +10,20 @@ find src -name component.ts | while read fn; do
     "chip")
       component="chips/chip"
       ;;
+    "chip-set")
+      component="chips/chip-set"
+      ;;
     *)
       ;;
   esac
-  cp material-components-web/packages/mdc-$component/component.ts $dir/component.ts.orig
-  diff -rupN $dir/component.ts.orig $dir/component.ts > $dir/component.patch || true
-  rm $dir/component.ts.orig
+  if test -d $dir/foundation.ts; then
+    cp material-foundations-web/packages/mdc-$foundation/foundation.ts $dir/foundation.ts.orig
+    diff -rupN $dir/foundation.ts.orig $dir/foundation.ts > $dir/foundation.patch || true
+    rm $dir/foundation.ts.orig
+  fi
+  if test -d $dir/component.ts; then
+    cp material-components-web/packages/mdc-$component/component.ts $dir/component.ts.orig
+    diff -rupN $dir/component.ts.orig $dir/component.ts > $dir/component.patch || true
+    rm $dir/component.ts.orig
+  fi
 done
